Split test_tags.py into test_tags.py file, testing AlgorithmTag and DifficultyTag classes and test_tag_proposals.py file, testing AlgorithmTagProposal and DifficultyTagProposal classes.
